desbaianado o evento da message os cmd de NSFW ta topper e bya bye DB (por enquanto)

+++ Bot/Comandos/Utility.cs
@@ -50,6 +50,7 @@ namespace Bot.Nucleo.Modulos
             {
                 context.Channel.SendMessageAsync(embed: new EmbedBuilder()
                         .WithColor(Color.DarkPurple)
+                        .WithTitle($"Canal: {usr.VoiceChannel.Name}")
                         .WithDescription($"[clique aqui](https://discordapp.com/channels/{context.Guild.Id}/{usr.VoiceChannel.Id}) para poder compartilhar sua tela ou ligar sua webcam")
                 .Build());
             }
@@ -64,11 +65,10 @@ namespace Bot.Nucleo.Modulos
 
         public void emote(CommandContext context, object[] args)
         {
+            string[] comando = (string[])args[1];
             try
             {
-                string[] comando = (string[])args[1];
                 Emote emote = Emote.Parse(comando[1]);
-
                 context.Channel.SendMessageAsync(embed: new EmbedBuilder()
                          .WithTitle(emote.Name)
                          .WithUrl(emote.Url)
@@ -76,41 +76,60 @@ namespace Bot.Nucleo.Modulos
                          .WithColor(Color.DarkPurple)
                      .Build());
             }
-            catch
+            catch (ArgumentException)
+            {
+
+                context.Channel.SendMessageAsync($"favor enfiar: {comando[1]} no cu do meu dono, agradecida");
+            }
+            catch (Exception e)
             {
                 context.Channel.SendMessageAsync(embed: new EmbedBuilder()
-                        .WithDescription($"**{context.User}** o emoji que você tentou usar é inválido")
-                        .AddField("Uso do comando: ", $"`{(string)args[0]}emote emoji`")
-                        .AddField("Exemplo: ", $"`{(string)args[0]}emote :kanna:`")
-                        .WithColor(Color.Red)
-                    .Build());
+                    .WithDescription($"**{context.User}** o emoji que você tentou usar é inválido {e}")
+                    .AddField("Uso do comando: ", $"`{(string)args[0]}emote emoji`")
+                    .AddField("Exemplo: ", $"`{(string)args[0]}emote :kanna:`")
+                    .WithColor(Color.Red)
+                .Build());
             }
+
         }
 
         public void say(CommandContext context, object[] args)
-        {
-
-            string[] comando = (string[])args[1];
-            string msg = "";
-            for (int i = 1; i < comando.Length; i++)
+        { 
+            if(!context.IsPrivate)
             {
-                msg += comando[i] + " ";
-            }
+                string[] comando = (string[])args[1];
+                string msg = "";
+                for (int i = 1; i < comando.Length; i++)
+                {
+                    msg += comando[i] + " ";
+                }
 
-            if (msg != "")
-            {
-                context.Message.DeleteAsync().GetAwaiter().GetResult();
+                if (msg != "")
+                {
+                    IGuildUser user = context.User as IGuildUser;
+                    if (user.GuildPermissions.ManageMessages)
+                    {
+                        context.Message.DeleteAsync().GetAwaiter().GetResult();
+                    }
 
-                context.Channel.SendMessageAsync(embed: new EmbedBuilder()
-                        .WithDescription(msg)
-                        .WithColor(Color.DarkPurple)
-                    .Build());
+                    context.Channel.SendMessageAsync(embed: new EmbedBuilder()
+                            .WithDescription(msg)
+                            .WithColor(Color.DarkPurple)
+                        .Build());
+                }
+                else
+                {
+                    context.Channel.SendMessageAsync(embed: new EmbedBuilder()
+                            .WithTitle("você precisa de me falar uma mensagem")
+                            .AddField("Uso do comando:", $"`{(string)args[0]}say <mensagem>`")
+                            .AddField("Uso do comando:", $"`{(string)args[0]}say @Thhrag#2527 sai do facebook`")
+                            .WithColor(Color.Red)
+                        .Build());
+                }
             } else
             {
                 context.Channel.SendMessageAsync(embed: new EmbedBuilder()
-                        .WithTitle("você precisa de me falar uma mensagem")
-                        .AddField("Uso do comando:", $"`{(string)args[0]}say <mensagem>`")
-                        .AddField("Uso do comando:", $"`{(string)args[0]}say @Thhrag#2527 sai do facebook`")
+                        .WithDescription("você so pode usar esse comando em servidores")
                         .WithColor(Color.Red)
                     .Build());
             }
@@ -119,13 +138,47 @@ namespace Bot.Nucleo.Modulos
         //nome bem top
         public void simg(CommandContext context, object[] args)
         {
-            string url = $"{context.Guild.IconUrl}?size=2048";
-            context.Channel.SendMessageAsync(embed: new EmbedBuilder()
-                    .WithTitle(context.Guild.Name)
-                    .WithDescription($"[Link Direto]({url})")
-                    .WithImageUrl(url)
-                    .WithColor(Color.DarkPurple)
-                .Build());
+            if(!context.IsPrivate)
+            {
+                string url = $"{context.Guild.IconUrl}?size=2048";
+                context.Channel.SendMessageAsync(embed: new EmbedBuilder()
+                        .WithTitle(context.Guild.Name)
+                        .WithDescription($"[Link Direto]({url})")
+                        .WithImageUrl(url)
+                        .WithColor(Color.DarkPurple)
+                    .Build());
+            } else
+            {
+                context.Channel.SendMessageAsync(embed: new EmbedBuilder()
+                        .WithDescription($"**{context.User}** esse comando so pode ser usado em servidores")
+                        .WithColor(Color.Red)
+                    .Build());
+            }
+        }
+
+        public void bigtext(CommandContext context, object[] args)
+        {
+            string[] comando = (string[])args[1];
+            string msg = "";
+            for (int i = 1; i < comando.Length; i++)
+            {
+                msg += comando[i] + " ";
+            }
+
+            char[] aa = msg.ToCharArray();
+
+            string txt = "";
+            for(int i=0; i< aa.Length; i++)
+            {
+                if (aa[i].ToString() == " ")
+                {
+                    txt += " ";
+                } else
+                {
+                    txt += $":regional_indicator_{aa[i]}:";
+                }
+            }
+            context.Channel.SendMessageAsync(txt.Substring(0, txt.Length - 22));
         }
     }
 }
